Affine params (#310)
This PR makes the checked Pedersen module take its parameters in affine coordinates so it doesn't spend time converting them to affine on the fly (which is super costly).

I also added "map" functions for tuples because I found myself wanting them when fixing up coda to build against this.