Fix C++ warnings (#360)
* Fix unsigned/signed comparison warnings

* Omit argument names for unused arguments in no_print

* Remove parameter from camlsnark_mnt*_fqk_one

This parameter should never have been there, and we got very lucky with
the calling convention: the parameter is compiled to be passed in a
register, and the register was never interfered with because the
argument was unused. If the argument had been stack-allocated, this
would have destroyed the stack.

Notably, this would have also caused us linking problems if we were
using Ctypes to generate the FFI statically for this function, instead
of using libffi at runtime.

* Remove unused assignment